LLM/requirements.txt
# Core Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
typing_extensions>=4.5.0  # Python 3.8 compatibility

# Deep Learning
torch>=2.0.0
# tensorflow>=2.13.0  # Optional: Commented out due to DLL issues on some systems

# LLM and NLP
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
langgraph>=0.0.30
langchain-core>=0.1.0
transformers>=4.35.0
sentence-transformers>=2.2.0
spacy>=3.7.0
nltk>=3.8.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
onnx>=1.15.0
onnxruntime>=1.16.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# Vector Databases
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pydantic>=2.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0
dask[complete]>=2023.12.0
distributed>=2023.12.0

# Logging and Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.26.0
ipywidgets>=8.1.0

# Additional NLP
textblob>=0.17.0
keybert>=0.8.0
